# Conflicts:
#	g202/SceneGame.cpp
#	g202/_BaseProject2019/Project/SceneTutorial.cpp
#	g202/_BaseProject2019/Project/WaterFlow.cpp

+++ g202/_BaseProject2019/Project/Player.h
@@ -275,7 +275,7 @@ public:
 		return causeOfDeath;
 	}
 	//進んだ距離(m数)を返す
-	int GetDistance()
+	double GetDistance()
 	{
 		//X座標の20分の1(適当)を進んだ距離とする
 		return ((posX - STARTPOS_X) / TRANSLATE_DISTANCE);
